Removed MultiMask (MaskMultiple). Added new pipeline stage MaskModel
