+++ src/main/java/org/openpnp/vision/pipeline/ui/CvPipelineEditor.java
@@ -38,11 +38,11 @@ import org.openpnp.vision.pipeline.stages.ImageWrite;
 import org.openpnp.vision.pipeline.stages.ImageWriteDebug;
 import org.openpnp.vision.pipeline.stages.MaskCircle;
 import org.openpnp.vision.pipeline.stages.MaskHsv;
+import org.openpnp.vision.pipeline.stages.MaskModel;
 import org.openpnp.vision.pipeline.stages.MaskRectangle;
 import org.openpnp.vision.pipeline.stages.MatchTemplate;
 import org.openpnp.vision.pipeline.stages.MinAreaRect;
 import org.openpnp.vision.pipeline.stages.MinAreaRectContours;
-import org.openpnp.vision.pipeline.stages.MultiMask;
 import org.openpnp.vision.pipeline.stages.Normalize;
 import org.openpnp.vision.pipeline.stages.ReadModelProperty;
 import org.openpnp.vision.pipeline.stages.Rotate;
@@ -93,11 +93,11 @@ public class CvPipelineEditor extends JPanel {
         registerStageClass(ImageWriteDebug.class);
         registerStageClass(MaskCircle.class);
         registerStageClass(MaskHsv.class);
+        registerStageClass(MaskModel.class);
         registerStageClass(MaskRectangle.class);
         registerStageClass(MatchTemplate.class);
         registerStageClass(MinAreaRect.class);
         registerStageClass(MinAreaRectContours.class);
-        registerStageClass(MultiMask.class);
         registerStageClass(Normalize.class);
         registerStageClass(ReadModelProperty.class);
         registerStageClass(Rotate.class);
